moved publish logic to headless listener, abstracted created document callback
