+++ src/main/java/edu/tamu/app/observer/AbstractDocumentListener.java
@@ -1,32 +1,25 @@
 package edu.tamu.app.observer;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import org.apache.commons.io.monitor.FileAlterationObserver;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 
 import edu.tamu.app.exception.DocumentNotFoundException;
 import edu.tamu.app.model.Document;
 import edu.tamu.app.model.IngestType;
 import edu.tamu.app.model.Project;
-import edu.tamu.app.model.ProjectRepository;
 import edu.tamu.app.model.repo.DocumentRepo;
 import edu.tamu.app.model.repo.ProjectRepo;
 import edu.tamu.app.service.DocumentFactory;
-import edu.tamu.app.service.registry.MagpieServiceRegistry;
-import edu.tamu.app.service.repository.Repository;
 
 public abstract class AbstractDocumentListener extends AbstractFileListener {
 
@@ -36,13 +29,6 @@ public abstract class AbstractDocumentListener extends AbstractFileListener {
 
     protected static final ExecutorService executor = Executors.newFixedThreadPool(10);
 
-    private static final List<Document> publishQueue = new CopyOnWriteArrayList<Document>();
-
-    private static final AtomicInteger publishing = new AtomicInteger(0);
-
-    @Value("${app.document.publish.concurrency:5}")
-    private int publishConcurrency;
-
     @Autowired
     private ProjectRepo projectRepo;
 
@@ -51,9 +37,6 @@ public abstract class AbstractDocumentListener extends AbstractFileListener {
 
     @Autowired
     protected DocumentRepo documentRepo;
-
-    @Autowired
-    private MagpieServiceRegistry projectServiceRegistry;
 
     public AbstractDocumentListener(String root, String folder) {
         super(root, folder);
@@ -69,16 +52,10 @@ public abstract class AbstractDocumentListener extends AbstractFileListener {
         Document existingDocument = documentRepo.findByProjectNameAndName(directory.getParentFile().getName(), directory.getName());
         if (existingDocument == null) {
             initializePendingResources(directory.getName());
-            createDocument(directory).thenAccept(newDocument -> {
-                if (newDocument != null) {
-                    logger.info("Document created: " + newDocument.getName());
-                    publishToRepositories(newDocument);
-                } else {
-                    logger.warn("Unable to create document!");
-                }
+            createDocument(directory).thenAccept(document -> {
+                createdDocumentCallback(document);
             });
         }
-
     }
 
     @Override
@@ -155,42 +132,16 @@ public abstract class AbstractDocumentListener extends AbstractFileListener {
         }, executor);
     }
 
-    protected void addResource(File file) throws DocumentNotFoundException {
-        documentFactory.addResource(file);
+    protected void createdDocumentCallback(Document document) {
+        if (document != null) {
+            logger.info("Document created: " + document.getName());
+        } else {
+            logger.warn("Unable to create document!");
+        }
     }
 
-    private void publishToRepositories(Document document) {
-
-        logger.info("Attempting to publish documnet: " + document.getName());
-
-        logger.info("Current concurrency: " + publishing.get());
-
-        if (publishing.get() <= publishConcurrency) {
-            publishing.incrementAndGet();
-
-            for (ProjectRepository repository : document.getProject().getRepositories()) {
-                try {
-                    document = ((Repository) projectServiceRegistry.getService(repository.getName())).push(document);
-                } catch (IOException e) {
-                    logger.error("Exception thrown attempting to push to " + repository.getName() + "!", e);
-                    e.printStackTrace();
-                }
-            }
-
-            publishing.decrementAndGet();
-
-            if (publishQueue.size() > 0) {
-                Document queuedDocument = publishQueue.get(0);
-                publishQueue.remove(0);
-                logger.info("Remaining in queue: " + publishQueue.size());
-                publishToRepositories(queuedDocument);
-            }
-
-        } else {
-            logger.info("Queueing document: " + document.getName());
-            publishQueue.add(document);
-        }
-
+    protected void addResource(File file) throws DocumentNotFoundException {
+        documentFactory.addResource(file);
     }
 
 }
